+++ core/src/main/java/com/cartoonsurvivors/game/entidades/Jugador.java
@@ -1,0 +1,8 @@
+package com.cartoonsurvivors.game.entidades;
+
+public class Jugador extends Entidad {
+    public Jugador(int vida, double posicionX, double posicionY) {
+        super(vida, posicionX, posicionY);
+    }
+
+}