+++ core/src/main/java/com/cartoonsurvivors/game/pantallas/PantallaJuego.java
@@ -1,0 +1,4 @@
+package com.cartoonsurvivors.game.pantallas;
+
+public class PantallaJuego {
+}